ml_service/requirements.txt
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
joblib==1.3.2

